+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - updateProfile function
 - updateTrack function
 - updateRepost function
+- setCategoryOnTrack function
 
 ### Removed
 - getGroupRatios function